requirements.common.txt
django==3.1
djangorestframework==3.14.0
django-environ==0.4.5
git+https://github.com/DES2048/django-vite
